+++ src/Symfony/Component/Validator/Constraints/SizeValidator.php
@@ -21,46 +21,6 @@ use Symfony\Component\Validator\ConstraintValidator;
  *
  * @deprecated Deprecated since version 2.1, to be removed in 2.3.
  */
-class SizeValidator extends ConstraintValidator
+class SizeValidator extends RangeValidator
 {
-    /**
-     * Checks if the passed value is valid.
-     *
-     * @param mixed      $value      The value that should be validated
-     * @param Constraint $constraint The constraint for the validation
-     *
-     * @return Boolean Whether or not the value is valid
-     *
-     * @api
-     */
-    public function validate($value, Constraint $constraint)
-    {
-        if (null === $value) {
-            return;
-        }
-
-        if (!is_numeric($value)) {
-            $this->context->addViolation($constraint->invalidMessage, array(
-                '{{ value }}' => $value,
-            ));
-
-            return;
-        }
-
-        if ($value > $constraint->max) {
-            $this->context->addViolation($constraint->maxMessage, array(
-                '{{ value }}' => $value,
-                '{{ limit }}' => $constraint->max,
-            ));
-
-            return;
-        }
-
-        if ($value < $constraint->min) {
-            $this->context->addViolation($constraint->minMessage, array(
-                '{{ value }}' => $value,
-                '{{ limit }}' => $constraint->min,
-            ));
-        }
-    }
 }
